Rename BOARD_STM32F4 to fix use with MB()
